moves map infowindow to another file
